extracted Searches into its own folder
