+++ we_are_vr/src/components/CustomAppBar.jsx
@@ -2,7 +2,7 @@ import AppBar from '@mui/material/AppBar';
 import Box from '@mui/material/Box';
 import Toolbar from '@mui/material/Toolbar';
 import Typography from '@mui/material/Typography';
-import SearchFunction from './SearchFunction';
+import SearchFunction from './searchFunctions/SearchFunction';
 import HomeIcon from '@mui/icons-material/Home';
 import PersonIcon from '@mui/icons-material/Person';
 import './AppBar.css';
