+++ we_are_vr/src/components/searchFunctions/SearchSubcriptions.jsx
@@ -2,31 +2,13 @@ import React, { useState } from 'react';
 import Autocomplete from '@mui/material/Autocomplete';
 import TextField from '@mui/material/TextField';
 import { useHistory } from 'react-router-dom';
-import allForums from '../Data/allForums';
+import subscribedForums from '../../Data/SubscribedForums';
+import handleForumClick from '../../utils/ForumNavigation';
 
-const SearchFunction = ({ barWidth }) => {
+const SearchSubcriptions = ( {barWidth} ) => {
     const [searchTerm, setSearchTerm] = useState('');
-    const [forums, setForums] = useState(allForums); // Set initial state to allForums
+    const [forums, setForums] = useState(subscribedForums);
     const history = useHistory();
-
-    const handleForumClick = (forumName) => {
-        const forumPaths = {
-            "v/Gorilla Tag": "/Forum",
-            "v/Quest 3": "/Home",
-            "v/Quest 2": "/Profile",
-            "v/GTA VI": "/Forum",
-            "v/GTA V": "/Forum",
-            "v/Spiderman": "/Forum",
-            "v/valve": "/Forum",
-            "v/MineCraft": "/Forum",
-            "v/Fortnite": "/Forum",
-            "v/Apex Legends": "/Forum",
-            "v/BeatSaber": "/Forum",
-            // Add the rest of the forums and their paths here
-        };
-
-        history.push(forumPaths[forumName]);
-    };
 
     const filteredForums = forums.filter(forum => forum.name.toLowerCase().includes(searchTerm.toLowerCase()));
 
@@ -38,14 +20,14 @@ const SearchFunction = ({ barWidth }) => {
             renderInput={(params) => (
                 <TextField
                     {...params}
-                    label="Search for forums"
+                    label="Subscribed forums"
                     variant="outlined"
                     sx={{backgroundColor: "white", margin:"0 1rem", width:barWidth}}
                 />
             )}
             renderOption={(props, option) => {
                 return (
-                    <li {...props} onClick={() => handleForumClick(option)}>
+                    <li {...props} onClick={() => handleForumClick(option,history)}>
                         {option}
                     </li>
                 );
@@ -54,4 +36,4 @@ const SearchFunction = ({ barWidth }) => {
     );
 };
 
-export default SearchFunction;
+export default SearchSubcriptions;
